+++ sdk/nodejs/bin/pulumi-provider-testing
