+++ sdk/nodejs/tsconfig.json
@@ -41,7 +41,7 @@
         "cmd/run/index.ts",
         "cmd/test-provider/index.ts",
 
-        "crud/index.ts",
+        "testing/index.ts",
 
         "tests/config.spec.ts",
         "tests/init.spec.ts",
