+++ sdk/proto/provider.proto
@@ -74,7 +74,7 @@ message CreateRequest {
 }
 
 message CreateResponse {
-    string id = 1;                         // the ID of the resource created.
+    string id = 1;                         // the ID of the created resource.
     google.protobuf.Struct properties = 2; // any properties that were computed during creation.
 }
